Merge EHU NLP and Readers

+++ metric-platform/platform-extensions/org.eclipse.scava.platform.communicationchannel.nntp.local/src/org/eclipse/scava/platform/communicationchannel/nntp/local/model/ArticleData.java
@@ -52,11 +52,11 @@ public class ArticleData extends Pongo {
 		notifyChanged();
 		return this;
 	}
-	public int getArticleNumber() {
-		return parseInteger(dbObject.get("articleNumber")+"", 0);
+	public long getArticleNumber() {
+		return parseLong(dbObject.get("articleNumber")+"", 0);
 	}
 	
-	public ArticleData setArticleNumber(int articleNumber) {
+	public ArticleData setArticleNumber(long articleNumber) {
 		dbObject.put("articleNumber", articleNumber);
 		notifyChanged();
 		return this;
